split into separate file
this all looks sorta necessary actually stripped some useless bits out
